% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseOps.R
\name{create_db_connection}
\alias{create_db_connection}
\title{Create a Database Connection}
\usage{
create_db_connection(user_name, password)
}
\arguments{
\item{user_name}{A string specifying the database username.}

\item{password}{A string specifying the database password.}
}
\value{
A database connection object (`DBI::dbConnect`) if the connection is successful.
        If the connection fails, returns NULL and prints an error message.
}
\description{
This function establishes a connection to an Azure SQL Server database
using the provided user credentials.
}
\examples{
# Create a database connection
con <- create_db_connection("your_username", "your_password")

# Use the connection to interact with the database
if (!is.null(con)) {
  dbListTables(con)  # List available tables
  dbDisconnect(con)  # Close the connection when done
}

}
