% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseOps.R
\name{create_table}
\alias{create_table}
\title{Create or Update a Table in a Database}
\usage{
create_table(connection, table_name, table, DATE_INSERTED = TRUE)
}
\arguments{
\item{connection}{A valid database connection object.}

\item{table_name}{A string specifying the name of the table to be created or updated.}

\item{table}{A data frame defining the structure of the table. If the data frame contains rows, they will be inserted into the database.}

\item{DATE_INSERTED}{A logical value (default: TRUE). If TRUE, ensures a `DATE_INSERTED` column exists in the database
and updates all its values to the current timestamp before insertion.}
}
\value{
NULL. The function creates or updates a table and prints confirmation messages.
}
\description{
This function creates a new table in a database based on the structure of a provided data frame.
If the table already exists, the user is prompted to either append or replace the data.
}
\examples{
# Create a new table with data
df <- data.frame(ID = 1:5, Name = c("A", "B", "C", "D", "E"))
create_table(my_connection, table_name = "NewTable", table = df)

# Create an empty table
empty_df <- data.frame(ID = integer(), Name = character())
create_table(my_connection, table_name = "EmptyTable", table = empty_df)

# Append or replace data in an existing table
new_data <- data.frame(ID = 6:10, Name = c("F", "G", "H", "I", "J"))
create_table(my_connection, table_name = "NewTable", table = new_data)

}
